Se agregaron los metodos de eliminar, actualizar, y mostrar en la capa de Logica de Negocios para la tabla Programa Educativo. En la capa de presentacio se hizo el diseño para el formulario RegistroProgramaEducativo,el cual ya elimina e inserta y muestra sus datos.

+++ ClassCapaEntidades/Carrera.cs
@@ -8,7 +8,7 @@ namespace ClassCapaEntidades
 {
     public class Carrera
     {
-        public int Id_Carrera { get; set; }
+        public byte Id_Carrera { get; set; }
         public string NombreCarrera { get; set; }
 
     }
